Improved error handling
The error handling has been improved by using a XML parser to get the
error codes instead of a regex. This makes it a lot more robust.

Additionally, the error message is also extracted from the response
if available. The error code is now used as the message_key, while the
error message is used as the human readable error message.
Finally, there is now special support for malformed XML responses,
allowing those to be caught separately if necessary.

Fixes #2.